+++ FEXCore/Source/Interface/Core/JIT/Arm64Relocations.cpp
@@ -6,7 +6,7 @@ desc: relocation logic of the arm64 splatter backend
 $end_info$
 */
 #include "Interface/Context/Context.h"
-#include "Interface/Core/JIT/Arm64/JITClass.h"
+#include "Interface/Core/JIT/JITClass.h"
 
 #include <FEXCore/Core/Thunks.h>
