+++ FEXCore/Source/Interface/Core/JIT/ConversionOps.cpp
@@ -5,7 +5,7 @@ tags: backend|arm64
 $end_info$
 */
 
-#include "Interface/Core/JIT/Arm64/JITClass.h"
+#include "Interface/Core/JIT/JITClass.h"
 
 namespace FEXCore::CPU {
 #define DEF_OP(x) void Arm64JITCore::Op_##x(IR::IROp_Header const* IROp, IR::NodeID Node)
