+++ FEXCore/Source/Interface/Core/JIT/JIT.cpp
@@ -16,7 +16,7 @@ $end_info$
 #include "Interface/Core/LookupCache.h"
 
 #include "Interface/Core/Dispatcher/Dispatcher.h"
-#include "Interface/Core/JIT/Arm64/JITClass.h"
+#include "Interface/Core/JIT/JITClass.h"
 
 #include "Interface/IR/Passes/RegisterAllocationPass.h"
